Experiment with a few [BET] Questions
The extension for CustomStringConvertible was not necessary. The initial structures can be defined as conforming to CustomStringConvertible. I haven't been able to tell if this is some type of "best practice" or an over-eager author on the web.

As expected the Set insert method handles duplicate inserts fine. I'm not sure there is a benefit to checking before doing the insert.